Implement imageAllowList settings and support them in the API
This commit does not provide the full functionality for `locked`
yet; it will be added once admin-provided allow lists are supported.

This commit creates the top-level object `containerEngine` to stop
putting container engine settings under `kubernetes`. Any settings
that are still applicable when kubernetes is disabled do not belong
in the `kubernetes` settings object.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>